refactoring: added BrowserRule class, common steps are moved to the BrowserRule
